Update RnnDbscan to build knn index
The classes have been updated according to the refactoring done on the
knn graph functions. They now build a knn index, which only needs to be
done once at construction time, then k can change without rebuilding the
index, which is the most expensive operation.

Additionally, Labels was changed to int32 to be more data-efficient,
especially when running parameter searches in a parfor where the labels
need to be sent back to the caller process.